File Open fails if the filename is not in English
HeifImage::init used CP_ACP to convert the Unicode filename from the File Open dialog to MBCS. But a filename with Unicode characters not mappable to extended ASCII in the current code page ended up getting corrupted, and caused heif_context_read_from_file to fail. So, this fix switches to CP_UTF8 so that any non-English Unicode characters are preserved and correctly passed to the heif read function.